ProjectArchive/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cssselect2==0.7.0
Django==5.0.3
et-xmlfile==1.1.0
fonttools==4.52.4
google-api-core==2.19.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
html5lib==1.1
httplib2==0.22.0
idna==3.7
lorem==0.1.1
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.2
pandas==2.2.1
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydyf==0.10.0
pyparsing==3.1.2
pyphen==0.15.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tinycss2==1.3.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
weasyprint==62.1
webencodings==0.5.1
xlwt==1.3.0
zopfli==0.2.3
